RPi/requirements.txt
dbus-python==1.3.2
lgpio==0.2.2.0
pycairo==1.26.0
PyGObject==3.48.2
rpi-lgpio==0.6
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
smbus==1.1.post2
